requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
alembic=1.4.1=pypi_0
antlr4-python3-runtime=4.8=pypi_0
azure-core=1.10.0=pypi_0
azure-storage-blob=12.7.1=pypi_0
blas=1.0=mkl
ca-certificates=2021.1.19=h06a4308_0
certifi=2020.12.5=py38h06a4308_0
cffi=1.14.4=pypi_0
chardet=4.0.0=pypi_0
click=7.1.2=pypi_0
cloudpickle=1.6.0=pypi_0
cryptography=3.3.1=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
databricks-cli=0.14.1=pypi_0
docker=4.4.1=pypi_0
entrypoints=0.3=pypi_0
flask=1.1.2=pypi_0
freetype=2.10.4=h5ab3b9f_0
gitdb=4.0.5=pypi_0
gitpython=3.1.12=pypi_0
gunicorn=20.0.4=pypi_0
hydra-core=1.0.6=pypi_0
idna=2.10=pypi_0
importlib-resources=5.1.0=pypi_0
intel-openmp=2020.2=254
isodate=0.6.0=pypi_0
itsdangerous=1.1.0=pypi_0
jinja2=2.11.3=pypi_0
jpeg=9b=h024ee3a_2
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuv=1.40.0=h7b6447c_0
lz4-c=1.9.3=h2531618_0
mako=1.1.4=pypi_0
markupsafe=1.1.1=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.2.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
mlflow=1.13.1=pypi_0
msrest=0.6.21=pypi_0
mypy=0.812=pyhd3eb1b0_0
mypy_extensions=0.4.3=py38_0
ncurses=6.2=he6710b0_1
ninja=1.10.2=py38hff7bd54_0
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
omegaconf=2.0.6=pypi_0
openssl=1.1.1j=h27cfd23_0
pandas=1.2.1=pypi_0
pillow=8.1.0=py38he98fc37_0
pip=20.3.3=py38h06a4308_0
prometheus-client=0.9.0=pypi_0
prometheus-flask-exporter=0.18.1=pypi_0
protobuf=3.14.0=pypi_0
psutil=5.8.0=py38h27cfd23_1
pycparser=2.20=pypi_0
python=3.8.5=h7579374_1
python-dateutil=2.8.1=pypi_0
python-editor=1.0.4=pypi_0
pytorch=1.7.1=py3.8_cuda10.2.89_cudnn7.6.5_0
pytorch-pretrained-vit=0.0.7=pypi_0
pytz=2021.1=pypi_0
pyyaml=5.4.1=pypi_0
querystring-parser=1.2.4=pypi_0
readline=8.1=h27cfd23_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
scipy=1.6.1=py38h91f5cce_0
setuptools=52.0.0=py38h06a4308_0
six=1.15.0=py38h06a4308_0
smmap=3.0.5=pypi_0
sqlalchemy=1.3.23=pypi_0
sqlite=3.33.0=h62c20be_0
sqlparse=0.4.1=pypi_0
tabulate=0.8.7=pypi_0
tk=8.6.10=hbc83047_0
torchaudio=0.7.2=py38
torchvision=0.8.2=py38_cu102
tqdm=4.56.0=pypi_0
typed-ast=1.4.2=py38h27cfd23_1
typing_extensions=3.7.4.3=pyh06a4308_0
urllib3=1.26.3=pypi_0
websocket-client=0.57.0=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0